Start tests for the instructions component
